[audio] AudioDeviceRegistry assembly input bundle
AudioDeviceRegistry, previously a static lib, is now a
standalone service binary and platform AIB. The media
config_schema has been created and includes a boolean
that determines on a per-product basis whether this
AIB is enabled.

If enabled, the service eagerly starts at bootup. This
is temporary while we complete the other layers of the
audio stack; eventually, it will be demand-started by
the first client connection.

Bug: 87650

Change-Id: Ib802c72581742caa17ca6c5f06de9c9ddb62a8a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738634
Reviewed-by: Cam Lloyd <camlloyd@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>