[fxfs] Fixes format error with full seek blocks.
Previously we would `skip()` forward the writer to the end of the block
but when the block was already full we'd accidentally add an extra
block. It isn't necessary to pad out to the end of the block, so just
stop doing that.

Also the WriteBytes trait has been clarified to always append zeroes
when calling `skip()` and the Writer struct has been updated accordingly
to align the behaviours of all implementations.

Bug: 127676

Change-Id: I4673ac434ed9a84a55d5761df4ab28346abb9f7f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863236
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>