[superproject] Roll llvm-project/libc [libc][NFC] Move the Linux file implementation to a subdirectory
This patch simply moves the special handling for `linux` files to a
subdirectory. This is done to make it easier in the future to extend
this support to targets (like the GPU) that will have different
dependencies.

Reviewed By: lntue

Differential Revision: https://reviews.llvm.org/D151231

GitOrigin-RevId: ec97744104b9804ec8ebb96c9379695ed2147c76
Original-Revision: 0198d3137188a200e465a1d43c18f376e2d048fd
Roller-URL: https://ci.chromium.org/b/8779666370050252241
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I820b329ecbf3b63240f932d9983096cc965f7a85
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863318