[fxfs] Unwrap keys on the pager executor
The main executor can make blocking syscalls to read the contents of a
file that reenter through the pager executor. The pager executor needs
the unwrapped keys to decrypt the file contents. If the keys are
unwrapped on the main executor then fxfs can deadlock. Moving the
unwrapping of the keys to the pager executor prevents the deadlock.

Fixed: 127056
Change-Id: Ib1b5acc2e59a88ad088a4cfd5fadbfeaf166d925
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862041
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Chris Suter <csuter@google.com>