[dart][log] Remove streams library
All this is unused at the moment. We can bring it back when we migrate
the dart library to encode the structured format.

Linking fxb/119265 so we can easily refer to this CL when we migrate
dart to use this encoding. Most of the work is probably already done
and should be an easy migration. We can also bring back the validator
when we actually use this code.

For now deleting to facilitate removing dart from fuchsia.git.

Fixed: b/240154194
Fixed: b/240154095
Bug: 119265

Change-Id: I8b712d39376c2ccfc5f5c6fe15b45ef1d4971d03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862616
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>