[superproject] Roll fuchsia [dart][log] Remove streams library
All this is unused at the moment. We can bring it back when we migrate
the dart library to encode the structured format.

Linking fxb/119265 so we can easily refer to this CL when we migrate
dart to use this encoding. Most of the work is probably already done
and should be an easy migration. We can also bring back the validator
when we actually use this code.

For now deleting to facilitate removing dart from fuchsia.git.

Original-Fixed: b/240154194
Original-Fixed: b/240154095
Original-Bug: 119265

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862616
Original-Revision: cb704d7910572c72eab5b7e70261b3bb68a43cb3
GitOrigin-RevId: 14d2f48b30ff3abfca4428c54925582b975c3ed4
Roller-URL: https://ci.chromium.org/b/8779625788647017265
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ifeee3b94b6d1ce6f6e748b92ed697152d9170682
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863876