[fdf][shutdown] Fix DFv1 ShutdownAllDrivers
Currently when the coordinator either:
 finishes shutting down all the drivers
 or encounters an error during shutdown,
it calls exit(0), which causes the kernel to
reboot because the driver manager is marked as a critical process.
This behavior is incorrect; it should instead
notify the shutdown manager via the callback that
it is finished shutting down drivers.
This allows the shutdown manager to shutdown
using the appropriate method.

Also removed unused suspend callback in main.cc

Bug: 125639, 125784

Testing: manual testing with ffx target reboot -b
Change-Id: I842eebf52e3d52ca96a5ff1fd4a0acfe10e406b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862823
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Garratt Gallagher <garratt@google.com>