[bazel][assembly] Use final_zbi in PB metadata
This is a followup on https://fxrev.dev/863156, which might be breaking
a few downstream builders.

There are builder logs suggesting `files.zbi` missing for unknown
reasons. For now, use final_zbi to match previous behavior for GN
assembled PBs, and it is also correctly set for Bazel assembled PBs.

Basically, this should continue to keep `ffx emu` working, and
potentially fix the problems are seeing in downstream builders. The
downside is that we have to live with a unnecessary copy for longer.

Bug: 127949

Change-Id: I4312ddfc81dc9fdde2b61fb236bc20dfb0954501
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863796
Fuchsia-Auto-Submit: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Chong Gu <chonggu@google.com>
Commit-Queue: Chong Gu <chonggu@google.com>