[superproject] Roll chromium/deps/icu@default Add language names for Translate
The following 11 languages are used by Translate so need to be added
to the Accept-Language list:

  ay, bho, bm, doi, dv, ilo, lus, mai, mni, sa, ts

"as" and "kok" will also be added as Translate languages, but are already supported by ICU.

We only need the language name from ICU in all UI languages and the
native name of each of the languages.

The following files are added in source/data/lang to add the native
display name of {ay,dv,ilo,lus,ts} along with the patch file and
a note about them in README.chromium.

ay.txt
dv.txt
ilo.txt
lus.txt
ts.txt

The data size difference is as follows:

         Old      New      Delta
chromeos 11718496 11730784 12288
common   10560960 10573104 12144

Original-Bug: 1313213
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromium/deps/icu/+/4563555
Original-Revision: 709d56e26482982f7390e49f98ef7a91101f137e
GitOrigin-RevId: b8c49b587b69e2e67197bc46d85c116f030d9179
Roller-URL: https://ci.chromium.org/b/8779619319107185825
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I62b31e933f63f8c349f5c8d1642e9af9a1212b3f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863078