[netstack3] Remove ICMP socket support in bindings
Remove bindings support for ICMP sockets. This enables the UDP interface
between core and bindings to be changed without having to make
equivalent but meaningless changes to the ICMP interface. As part of
this change, mark previously-passing tests that relied on ICMP sockets
as expected to fail.

This represents a regression in behavior for Netstack3 from the point
of view of passing tests. That's okay; the ICMP socket implementation
is scheduled for an overhaul anyway, and was largely unused (and not
load-bearing). In return, the changes required to merge UDP socket IDs
are significantly simplified by not having to deal with ICMP sockets.
Once ICMP sockets are re-implemented in the Netstack3 core as datagram sockets (https://fxbug.dev/125482), they will be reintroduced in bindings.

Bug: 125482, 47321
Change-Id: I87b797744ff3cf6e3ee6cd16ee9a74f2ed07a448
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863677
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>