[pci] Allocate BARs with existing addresses first
On NUC11 with an external wifi chip it was found that since the
bootloader only configured BAR2 and provided it the address of the start
of our allocator it became possible for the allocation of BAR0 to
temporarily share the same address until BAR2 was reallocated. Rather
than temporarily disabling either BAR, this resulted in BAR0 not being
mapped at all, while BAR2 continued to operate as normal. To prevent
this corner case we'll allocate the BARs with preconfigured addresses
first.

This is confirmed to fix the wifi issue, but a test to replicate this
situation is still needed. This will delay the fix so it will be
tracked in fxb/128196

Bug: 128196

Change-Id: I06fc08f9af8e12778ce85fec5966379dd66618d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862362
Fuchsia-Auto-Submit: Christopher Anderson <cja@google.com>
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>