[starnix] Enable in-thread exception handling
This enables the in-thread exception handling path in Starnix by
default.

This has a subtle dependency on commit
612804c28cc8c0d123a00b83e4df316436e10e97 in that the thread that
generated the exception is also used for handling it so we must preserve
all architectural state observable by Linux. This adds a test that we
preserve the contents of the first vector register which is used in
several places on the Starnix side.

The threaded exception handling path and configuration is preserved
for now to make debugging issues with this codepath easier. Once the
in-thread exception flow proves stable we can remove the other path.

Fixed: 117302
Change-Id: I7f937c7632044f086884a8b124b91973badaac24
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864159
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>