[rust-3p] Import loom
This patch imports the library loom 0.5.6. In order to do this, it also:

* Adds generator 0.7.4
* Adds matchers 0.1.0
* Enables the regex-automata feature "regex-syntax"
* Enables the tracing-subscriber feature "env-filter"
* Adds a stub windows rustversion 1.0.12
* Adds scoped-tls 1.0.1
* Adds a stub windows module 0.48.0

loom, generator, matchers, and scoped-tls were approved in OSRB
http://fxbug.dev/107676.

Since the loom crate is intended to be only used in tests, it's marked
as `testonly`. Furthermore, the `generator` indirect dependency uses
assembly to implement green threads, so to reduce our risk of bugs with
it, it also is being marked as `testonly` and restricted to only be used
by `loom`.

Bug: 107676
Bug: 127528

Change-Id: Ibb30cabc7533f9413cecfe7a5fee946a6bce4e98
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718991
Reviewed-by: David Koloski <dkoloski@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>