[starnix] Distinguish inode and node identifier in FsNode
Before this CL, the FsNode was using the inode as the unique identifier
of the node, used by the filesystem for caching.

This is problematic for FS like fuse where node needs to be created
before their inode is known. Instead, FsNode now have an unique
identifier that is independent of their inode, but unless the FsNode
implementation changes the inode, it will by default, be set to the node
identifier.

Bug: 125496
Change-Id: I188a742a934c7884ba440a55aa5439d35109fcf9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864536
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>