[bt][gap] Add past peers + discovery to inspect
Add a limited list of peers that were recently removed to the inspect
tree, allowing debugging to notice if a peer was discovered and expired
from the cache.

Add names + addresses to peer inspect.

Add a list of recent discovery sessions and their durations, and a count
of total discovery sessions.

Fixed: 84985
Bug: b/282091832
Change-Id: I13ad02e3983887a5de0ae11e784f8e37d7ca7df9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862777
Commit-Queue: Marie Janssen <jamuraa@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>