[check-licenses] Add ability to use license classifier
licenseclassifier is an open-source tool that can accurately
detect license texts. It's detection algo is much better than the
strategy that check-licenses currently uses.

This CL adds the licenseclassifier library to golibs,
and hooks it up in check-licenses behind a flag. It will be turned
on in a future CL.

Bug: 106919, 127868
Change-Id: I6b2f3e42f10a757013d698f0025dcb40d0b71d0d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859884
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>