[check-licenses] Add license text catalog.
This library of license texts will be used by licenseclassifier
to detect license texts in the fuchsia tree. This should be faster
and more accurate than the current strategy that check-licenses
employs.

This just adds the license texts, the new library will be enabled
in a future CL.

Change-Id: I73c536c53428a0b34458a585fa662a51585295d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859494
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Jay Patel <pateljay@google.com>