[build] Use the unversioned ffx binaries for build actions
Build actions should not need the version-stamped ffx
binaries, and using the non-stamped binaries will reduce
the amount of the tree that is dirtied when changing the
git integration commit patch that is stamped into the
binaries.

The result should be fewer targets being needlessly
rebuilt for developers after a jiri update or in CI/CQ
due to the change in the information used to stamp the
binaries.

Change-Id: I36b552ed4de30eda45dd040ca47fc124a53bf239
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/865021
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>