[superproject] Roll llvm-project/libc [libc] Add strtoint32 and strtoint64 tests
There were regressions in the testing framework due to none of the
functioning buildbots having a 32 bit long. This allowed the 32 bit
version of the strtointeger function to go untested. This patch adds
tests for strtoint32 and strtoint64, which are internal testing
functions that use constant integer sizes. It also fixes the tests to
properly handle these situations.

Reviewed By: sivachandra

Differential Revision: https://reviews.llvm.org/D151935

GitOrigin-RevId: e5b2e8e7f4ccf44724e0c25588c438dfcc94d45b
Original-Revision: 0338bc48bf0854728ab3162cde2edb3364b154af
Roller-URL: https://ci.chromium.org/b/8779404853011047473
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ic64398726597599384ce1125e4c702d5aed7977a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/865197