[memory_sampler] Support dynamic memory allocation in hooks
Currently, allocations are forbidden from the scudo
allocation/deallocation hooks, (mostly) because of the risk of
infinite recursion. This commit adds a simple re-entrancy guard, which
is used to detect and prevent such a loop.

Being able to allocate from the hooks is desirable: the usual strategy
to handle deallocations in a sampling profiler is to store, at
allocation time, the sampled allocations in a dynamically allocated
set, and to consult this set at deallocation time to decide which
deallocations need to be reported.

Note that this use case is technically forbidden per Scudo's contract,
though it appears to work well enough, from our manual testing. We
are working with the Scudo owners to lift this restriction. Depending
on where this discussion lands, we might want to revisit this approach
in a subsequent commit.

Change-Id: I1bd9622b93158c26abec347627586060fdce7778
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864018
Reviewed-by: Étienne J. Membrives <etiennej@google.com>
Reviewed-by: Fabio D'Urso <fdurso@google.com>
Commit-Queue: Guillaume Chelfi <chelfi@google.com>