[concurrent][docs] Improve seqlock documentation.
Make an attempt to improve the documentation surrounding the proper
use of a libconcurrent SeqLock. Using a sequence lock correctly brings
with it some pretty specific requirements in order to make things work
correctly across all architectures, independent of any individual
architecture's memory model.

So, update the README to be a bit more "cookbook-like".  Give proper
warnings that this can be tricky stuff, and provide _specific_ ways
that users should always follow when using a SeqLock in their code.

In addition, given that this is tricky stuff, add some (separate)
documentation which talks about why we think that the current
implementation is guaranteed to work, and be portable across all
architectures.  In specific, add a `models` directory which contains:

++ A README which walks a reader through a set of arguments for why
   what we have here works.  Define "success", state the requirements
   for achieving success, and finally demonstrate how the choice of
   memory-order semantics guarantees that the requirements are met.
++ Examples of "good" synchronization patterns which achieve these
   requirements, written as models which can be checked using
   CDSChecker.  A "minimal" (single writer) sequence lock model is
   provided here, as are the two more complex (multi-writer) versions
   that we actually offer via the library.  One which uses Acq/Rel
   semantics on payload loads/stores, and another which uses relaxed
   ops for payloads, and fence-to-fence synchronization instead.
++ Finally, also include some models "bad" synchronization attempts.
   The idea here is that it can be easy to look at the existing
   algorithm and wonder, "why can't I just do this simpler thing
   instead?".  This has happened to me many times when working on this
   problem, so I thought it would be good to write down the bad ideas
   that I have had, and provide a write-up of exactly why they _don't_
   work.  Hopefully, if I (or someone else) ever has an idea like this
   in the future, then can come here and find a reasonable
   explanation of why the idea does not actually work like someone
   might think that it should.

Change-Id: If15362bff662fa2c1185fa5a0619ef1db474938c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/809939
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: John Grossman <johngro@google.com>