[remote-control] Add log API
Allow host-side tools to emit a log on the device, similar to the SL4F
logging facade.

Many product E2E tests make use of SL4F the logging facade to do this.
This device-side logging allows test authors to sequence and correlate
specific events in the test with events happening on the device, for
example to know in the device logs where the beginning and the end of a
test case was. That is not easy to do if the log only exist in the test's
output.

It would be a pretty significant regression if Lacewing-based tests
couldn't do the same thing.

Bug: b/276968393 b/276726514
Change-Id: I644b5f894e4877ad88adac60a9a422faf858e3ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862578
Reviewed-by: Andrew Davies <awdavies@google.com>
API-Review: Christopher Johnson <crjohns@google.com>
Commit-Queue: David Worsham <dworsham@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>