[story_runner] Pass lambda which accepts AnnotationController_WatchAnnotations_Result
UBSan was detecting a function type mismatch on the invoked callback
passed to GetAnnotations. The callback is a fit::function whose argument
expected to be a AnnotationController_GetAnnotations_Result, but the
underlying function that's actually called is a
AnnotationController_WatchAnnotations_Result. Structurally, both result
objects are the same. They're essentially structs that hold a vector of
Annotations.

Bug: 128274
Change-Id: I5de445b8e9f5025f8c645401bf675a97d38ccc2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/867705
Reviewed-by: Gabe Schine <thatguy@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>