[elfldltl] Use AbiPtr et al in Soname, SymbolInfo, InitFiniInfo
These data structures will be used in the new dynamic linker's
"passive" ABI, so define them to be compatible with the out of
process case.  They must be templated on not only the Elf layout
but on the AbiTraits to use in their pointer and span members.
This also includes their subordinate data structures such as
CompatHash and GnuHash.

These classes are still only fully usable in their LocalAbiTraits
instantiations.  The testing of other instantiations is limited
to verifying that the types can be instantiated with the stub
RemoteAbiTraits and still support default- and copy-construction
and copy-assignment.

Bug: 121817
Change-Id: I13a013b2df55b3b408bf7cc59ca76f0792d54e81
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/866118
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Alex Brachet <abrachet@google.com>